+++ src-dbg/flow_storm/debugger/ui/plugins.clj
@@ -1,6 +1,5 @@
 (ns flow-storm.debugger.ui.plugins
-  (:require [flow-storm.utils :as utils :refer [log-error log]]
-            [clojure.string :as str]))
+  (:require [flow-storm.utils :as utils :refer [log-error log]]))
 
 (defonce *plugins (atom {}))
 
@@ -27,10 +26,8 @@
 (defn plugins []
   (vals @*plugins))
 
-(defn load-plugins-namespaces [{:keys [plugins-ns-str]}]
-  (when-not (str/blank? plugins-ns-str)
-    (let [namespaces (->> (str/split plugins-ns-str #",")
-                          (mapv symbol))]
-      (doseq [n namespaces]
-        (log (format "Requiring plugin ns %s" n))
-        (require n)))))
+(defn load-plugins-namespaces [{:keys [plugins-namespaces-set]}]
+  (let [namespaces (mapv symbol plugins-namespaces-set)]
+    (doseq [n namespaces]
+      (log (format "Requiring plugin ns %s" n))
+      (require n))))